Extract CLI test fixtures
